# api/tm-catalog.openapi.yaml
openapi: 3.0.3
info:
  title: Thing Model Catalog - OpenAPI 3.0
  version: 0.1.2
  description: "This is the Thing Model Catalog API specification."
  contact: {}  
...
servers:
  - url: ''
security:
  - BearerAuth: []
...
paths:
  /inventory:
    get:
      tags:
        - inventory
      summary: Get the inventory of the catalog
      description: Returns the catalogs inventory
      operationId: getInventory
      parameters:
        - name: 'repo'
          in: query
          description: Source repository name. Optionally constrains the results to only those from given named repository. See '/repos'
          required: false
          schema:
            type: string
          example: 'global'
        - name: 'filter.author'
          in: query
          description: |
            Filters the inventory by one or more authors having exact match.  
            The filter works additive to other filters.
          schema:
            type: string
          example: 'MyCompany,siemens'
        - name: 'filter.manufacturer'
          in: query
          description: |
            Filters the inventory by one or more manufacturers having exact match.  
            The filter works additive to other filters.
          schema:
            type: string
          example: 'BarTech,siemens'
        - name: 'filter.mpn'
          in: query
          description: |
            Filters the inventory by one ore more mpn (manufacturer part number) having exact match.   
            The filter works additive to other filters.
          schema:
            type: string
          example: 'BazLamp,POC1000'
        - name: 'filter.name'
          in: query
          description: |
            Filters the inventory by inventory entry name having a prefix match of full path parts.   
            The filter works additive to other filters.
          schema:
            type: string
          example: 'siemens/siemens/poc1000'
        - name: 'search'
          in: query
          description: |
            Filters the inventory according to whether the content of the inventory entries matches the given search.    
            The search works additive to other filters.
          schema:
            type: string
          example: ''
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryResponse'
              examples:
                inventory:
                  $ref: '#/components/examples/InventoryResponseExample'
        '400':
          description: Invalid filter parameter supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Upstream repository error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /inventory/{tmID}:
    get:
      tags:
        - inventory
      summary: Get the metadata of a single TM by ID
      description: Returns the metadata of a single TM
      operationId: getInventoryByID
      parameters:
        - $ref: '#/components/parameters/TMID'
        - $ref: '#/components/parameters/Repo'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryEntryVersionResponse'
              examples:
                inventoryEntryVersion:
                  $ref: '#/components/examples/InventoryEntryVersionResponseExample'
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Inventory entry not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /inventory/.latest/{fetchName}:
    get:
      tags:
        - inventory
      summary: Get the metadata of the most recent TM version matching the name
      description: Returns the metadata of the most recent TM version matching the fetch name
      operationId: getInventoryByFetchName
      parameters:
        - $ref: '#/components/parameters/FetchName'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryEntryVersionResponse'
              examples:
                inventoryEntryVersion:
                  $ref: '#/components/examples/InventoryEntryVersionResponseExample'
        '400':
          description: Invalid fetch name supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Inventory entry not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /inventory/.tmName/{tmName}:
    get:
      tags:
        - inventory
      summary: Get an inventory entry by TM name
      description: Returns a single inventory entry
      operationId: getInventoryByName
      parameters:
        - name: tmName
          in: path
          description: Name of the inventory entry
          required: true
          schema:
            type: string
          example: 'siemens/siemens/poc1000'
        - $ref: '#/components/parameters/Repo'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryEntryResponse'
              examples:
                inventoryEntry:
                  $ref: '#/components/examples/InventoryEntryResponseExample'
        '400':
          description: Invalid TM name supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Inventory entry not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /thing-models/{tmID}:
    get:
      tags:
        - thing-models
      summary: Get the content of a Thing Model by its ID
      description: > 
        Returns the actual content of a Thing Model.
      operationId: getThingModelById
      parameters:
        - $ref: '#/components/parameters/TMID'
        - name: restoreId
          in: query
          description: restore the TM's original external id, if it had one
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: |
            Successful operation 

            **For the schema of the returned Thing Model see** [Thing Model JSON schema](https://github.com/w3c/wot-thing-description/blob/main/validation/tm-json-schema-validation.json)
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Invalid ID or fetch name supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Content of the Thing Model not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - thing-models
      summary: Delete a Thing Model by ID
      description: >
        Deletes a Thing Model.
        Normally, the catalog is write-only and no TMs are ever deleted.
        The delete operation is implemented for the rare cases when a TM has been imported whilst containing major errors 
        or by mistake. It is mandatory to provide the query parameter ?force=true to delete a TM.
      operationId: deleteThingModelById
      parameters:
        - $ref: '#/components/parameters/TMID'
        - name: force
          in: query
          description: flag to force the deletion. must be set to "true"
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Successful operation
        '400':
          description: Invalid ID or 'force' parameter supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Thing Model not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /thing-models/.latest/{fetchName}:
    get:
      tags:
        - thing-models
      summary: Get the content of a Thing Model by fetch name
      description: > 
        Returns the actual content of the Thing Model selected by given fetch name.
      operationId: getThingModelByFetchName
      parameters:
        - $ref: '#/components/parameters/FetchName'
        - name: restoreId
          in: query
          description: restore the TM's original external id, if it had one
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: |
            Successful operation 

            **For the schema of the returned Thing Model see** [Thing Model JSON schema](https://github.com/w3c/wot-thing-description/blob/main/validation/tm-json-schema-validation.json)
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Invalid ID or fetch name supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Content of the Thing Model not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /thing-models/{tmID}/.attachments/{attachmentFileName}:
    get:
      tags:
        - attachments
      summary: Get the actual content of an attachment to a Thing Model
      description:
        Returns the content of an attachment to a Thing Model.
        Use '/thing-models/.tmName/{tmName}/.attachments/{attachmentFileName}' endpoint to get TM name's attachments
      operationId: getThingModelAttachmentByName
      parameters:
        - $ref: '#/components/parameters/TMID'
        - $ref: '#/components/parameters/AttachmentFileName'
      responses:
        '200':
          description: Successful operation
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          description: Invalid ID requested
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: TM or attachment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - attachments
      summary: Upload an attachment to a Thing Model
      description: Upload an attachment to a Thing Model
      operationId: putThingModelAttachmentByName
      parameters:
        - $ref: '#/components/parameters/TMID'
        - $ref: '#/components/parameters/AttachmentFileName'
      requestBody:
        description: Add a new attachment or overwrite an existing one
        content:
          '*/*':
            schema:
              type: string
              format: binary
        required: true
      responses:
        '204':
          description: Successfully imported
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - attachments
      summary: Delete an attachment to a Thing Model
      description: Delete an attachment to a Thing Model
      operationId: deleteThingModelAttachmentByName
      parameters:
        - $ref: '#/components/parameters/TMID'
        - $ref: '#/components/parameters/AttachmentFileName'
      responses:
        '204':
          description: Successful operation
        '400':
          description: Invalid ID or inventory name supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Thing Model or inventory name or attachment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /thing-models/.tmName/{tmName}/.attachments/{attachmentFileName}:
    get:
      tags:
        - attachments
      summary: Get the actual content of an attachment to a TM name
      description:
        Returns the content of an attachment to a TM name
      operationId: getTMNameAttachment
      parameters:
        - $ref: '#/components/parameters/TMName'
        - $ref: '#/components/parameters/AttachmentFileName'
      responses:
        '200':
          description: Successful operation
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          description: Invalid TM name requested
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: TM name or attachment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - attachments
      summary: Upload an attachment to a TM name
      description: Upload an attachment to a TM name
      operationId: putTMNameAttachment
      parameters:
        - $ref: '#/components/parameters/TMName'
        - $ref: '#/components/parameters/AttachmentFileName'
      requestBody:
        description: Add a new attachment or overwrite an existing one
        content:
          '*/*':
            schema:
              type: string
              format: binary
        required: true
      responses:
        '204':
          description: Successfully added
        '400':
          description: Invalid TM name supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - attachments
      summary: Delete an attachment to a TM name
      description: Delete an attachment to a TM name
      operationId: deleteTMNameAttachment
      parameters:
        - $ref: '#/components/parameters/TMName'
        - $ref: '#/components/parameters/AttachmentFileName'
      responses:
        '204':
          description: Successful operation
        '400':
          description: Invalid TM name supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: TM name or attachment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /thing-models:
    post:
      tags:
        - thing-models
      summary: Import a Thing Model
      description: Import a Thing Model
      operationId: importThingModel
      parameters:
        - name: force
          in: query
          description: flag to force the import, ignoring any conflicts with existing TMs
          required: false
          schema:
            type: string
        - name: optPath
          in: query
          description: optional path parts to append to the target path (and id) of imported TM, after the mandatory path structure
          required: false
          schema:
            type: string
      requestBody:
        description: |
          Import a Thing Model  
          
          **For the schema of a Thing Model see** [Thing Model JSON schema](https://github.com/w3c/wot-thing-description/blob/main/validation/tm-json-schema-validation.json)
        content:
          application/json:
            schema:
              type: object
            examples:
              payload:
                $ref: '#/components/examples/ImportTMPayloadExample'
        required: true
      responses:
        '201':
          description: Successfully imported
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportThingModelResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          description: Conflict, Thing Model already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /repos:
    get:
      tags:
        - repos
      summary: Get the list of repositories
      description: |
        Get the list of repositories available for storing TMs. 
        When the list contains only one element, specifying the target repository for read and write operations is optional.
      operationId: getRepos
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReposResponse'
              examples:
                mpns:
                  $ref: '#/components/examples/ReposResponseExample'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    InventoryEntryVersionResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/InventoryEntryVersion'
...
    InventoryResponse:
      type: object
      required:
        - data
      properties:
        meta:
          $ref: '#/components/schemas/Meta'
        data:
          type: array
          items:
            $ref: '#/components/schemas/InventoryEntry'
    InventoryEntryResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/InventoryEntry'
    ImportThingModelResponse:
      required:
        - data
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ImportThingModelResult'
...
    ErrorResponse:
      required:
        - title
        - status
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
        status:
          type: integer
  parameters:
...
    TMID:
      name: tmID
      in: path
      description: TM ID
      required: true
      schema:
        type: string
      example: 'siemens/siemens/poc1000/v0.0.0-20231201133246-e1594d08a01b.tm.json'
...
    AttachmentFileName:
      name: attachmentFileName
      in: path
      description: File name of the attachment
      required: true
      schema:
        type: string
      example: 'README.md'
    Repo:
      name: repo
      in: query
      description: Source/target repository name. Required when repository is ambiguous. See '/repos'
      required: false
      schema:
        type: string
      example: 'global'
  responses:
    UnauthorizedError:
      description: API key is missing or invalid
      headers:
        WWW_Authenticate:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT